+++ lib/livebook/runtime/mix_standalone.ex
@@ -105,6 +105,7 @@ defmodule Livebook.Runtime.MixStandalone do
   defp start_elixir_mix_node(elixir_path, node_name, eval, argv, project_path) do
     # Here we create a port to start the system process in a non-blocking way.
     Port.open({:spawn_executable, elixir_path}, [
+      :binary,
       # We don't communicate with the system process via stdio,
       # contrarily, we want any non-captured output to go directly
       # to the terminal