+++ lib/livebook_web/controllers/health_controller.ex
@@ -2,11 +2,10 @@ defmodule LivebookWeb.HealthController do
   use LivebookWeb, :controller
 
   def index(conn, _params) do
-    version = Application.spec(:livebook, :vsn) |> List.to_string()
-
-    json(conn, %{
-      "application" => "livebook",
-      "version" => version
+    conn
+    |> put_resp_header("Access-Control-Allow-Origin", "*")
+    |> json(%{
+      "application" => "livebook"
     })
   end
 end