+++ lib/livebook_web/live/session_live/cell_component.ex
@@ -56,14 +56,24 @@ defmodule LivebookWeb.SessionLive.CellComponent do
     <div class="mb-1 flex items-center justify-between">
       <div class="relative z-20 flex items-center justify-end space-x-2" data-element="actions" data-primary>
         <%= if @cell_view.evaluation_status == :ready do %>
-          <button class="text-gray-600 hover:text-gray-800 focus:text-gray-800 flex space-x-1 items-center"
-            phx-click="queue_cell_evaluation"
-            phx-value-cell_id={@cell_view.id}>
-            <.remix_icon icon="play-circle-fill" class="text-xl" />
-            <span class="text-sm font-medium">
-              <%= if(@cell_view.validity_status == :evaluated, do: "Reevaluate", else: "Evaluate") %>
-            </span>
-          </button>
+          <%= if @cell_view.validity_status != :fresh and @cell_view.reevaluate_automatically do %>
+            <%= live_patch to: Routes.session_path(@socket, :cell_settings, @session_id, @cell_view.id),
+                  class: "text-gray-600 hover:text-gray-800 focus:text-gray-800 flex space-x-1 items-center" do %>
+              <.remix_icon icon="check-line" class="text-xl" />
+              <span class="text-sm font-medium">
+                Reevaluates automatically
+              </span>
+            <% end %>
+          <% else %>
+            <button class="text-gray-600 hover:text-gray-800 focus:text-gray-800 flex space-x-1 items-center"
+              phx-click="queue_cell_evaluation"
+              phx-value-cell_id={@cell_view.id}>
+              <.remix_icon icon="play-circle-fill" class="text-xl" />
+              <span class="text-sm font-medium">
+                <%= if(@cell_view.validity_status == :evaluated, do: "Reevaluate", else: "Evaluate") %>
+              </span>
+            </button>
+          <% end %>
         <% else %>
           <button class="text-gray-600 hover:text-gray-800 focus:text-gray-800 flex space-x-1 items-center"
             phx-click="cancel_cell_evaluation"
@@ -336,7 +346,7 @@ defmodule LivebookWeb.SessionLive.CellComponent do
 
   defp cell_status(%{cell_view: %{evaluation_status: :queued}} = assigns) do
     ~H"""
-    <.status_indicator circle_class="bg-gray-500" animated_circle_class="bg-gray-400">
+    <.status_indicator circle_class="bg-gray-400" animated_circle_class="bg-gray-300">
       Queued
     </.status_indicator>
     """
@@ -363,7 +373,7 @@ defmodule LivebookWeb.SessionLive.CellComponent do
 
   defp cell_status(%{cell_view: %{validity_status: :aborted}} = assigns) do
     ~H"""
-    <.status_indicator circle_class="bg-red-400">
+    <.status_indicator circle_class="bg-gray-500">
       Aborted
     </.status_indicator>
     """
@@ -419,12 +429,18 @@ defmodule LivebookWeb.SessionLive.CellComponent do
     <div class="flex flex-col rounded-lg border border-gray-200 divide-y divide-gray-200">
       <%= for {output, index} <- @cell_view.outputs |> Enum.reverse() |> Enum.with_index(), output != :ignored do %>
         <div class="p-4 max-w-full overflow-y-auto tiny-scrollbar">
-          <%= render_output(output, %{id: "cell-#{@cell_view.id}-evaluation#{@cell_view.number_of_evaluations}-output#{index}", socket: @socket}) %>
+          <%= render_output(output, %{
+                id: "cell-#{@cell_view.id}-evaluation#{evaluation_number(@cell_view.evaluation_status, @cell_view.number_of_evaluations)}-output#{index}",
+                socket: @socket
+              }) %>
         </div>
       <% end %>
     </div>
     """
   end
+
+  defp evaluation_number(:evaluating, number_of_evaluations), do: number_of_evaluations + 1
+  defp evaluation_number(_evaluation_status, number_of_evaluations), do: number_of_evaluations
 
   defp render_output(text, %{id: id}) when is_binary(text) do
     # Captured output usually has a trailing newline that we can ignore,